Rename Formatter to Generator
