+++ spec/Generator/UrlsetSpec.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace spec\Thepixeldeveloper\Sitemap\Formatter;
+namespace spec\Thepixeldeveloper\Sitemap\Generator;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
-use Thepixeldeveloper\Sitemap\Sitemap;
-use Thepixeldeveloper\Sitemap\SitemapIndex;
+use Thepixeldeveloper\Sitemap\Urlset;
 
-class SitemapIndexSpec extends ObjectBehavior
+class UrlsetSpec extends ObjectBehavior
 {
     function let(\XMLWriter $writer)
     {
@@ -16,19 +15,19 @@ class SitemapIndexSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType('Thepixeldeveloper\Sitemap\Formatter\SitemapIndex');
+        $this->shouldHaveType('Thepixeldeveloper\Sitemap\Generator\Urlset');
     }
 
-    function it_should_generate_the_right_sitemap_index_attributes(SitemapIndex $sitemapIndex, \XMLWriter $writer)
+    function it_should_generate_the_right_sitemap_index_attributes(Urlset $urlset, \XMLWriter $writer)
     {
-        $sitemapIndex->getSitemaps()->willReturn([]);
+        $urlset->getUrls()->willReturn([]);
 
-        $writer->startElement('sitemapindex')->shouldBeCalled();
+        $writer->startElement('urlset')->shouldBeCalled();
         $writer->writeAttribute('xmlns:xsi', 'http://www.w3.org/2001/XMLSchema-instance')->shouldBeCalled();
         $writer->writeAttribute('xsi:schemaLocation', 'http://www.sitemaps.org/schemas/sitemap/0.9 http://www.sitemaps.org/schemas/sitemap/0.9/siteindex.xsd')->shouldBeCalled();
         $writer->writeAttribute('xmlns', 'http://www.sitemaps.org/schemas/sitemap/0.9')->shouldBeCalled();
         $writer->endElement()->shouldBeCalled();
 
-        $this->generate($sitemapIndex);
+        $this->generate($urlset);
     }
 }
